refactor: "나의 예약" 관련 로직 분리
