+++ src/main/java/roomescape/presentation/api/MyReservationController.java
@@ -3,8 +3,7 @@ package roomescape.presentation.api;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RestController;
-import roomescape.application.MyReservationQueryService;
-import roomescape.application.ReservationService;
+import roomescape.application.MyReservationService;
 import roomescape.presentation.AuthenticationPrincipal;
 import roomescape.presentation.dto.request.LoginMember;
 import roomescape.presentation.dto.response.MyReservationResponse;
@@ -15,17 +14,15 @@ import java.util.List;
 @RestController
 public class MyReservationController {
 
-    private final ReservationService reservationService;
-    private final MyReservationQueryService reservationQueryService;
+    private final MyReservationService reservationQueryService;
 
-    public MyReservationController(ReservationService reservationService, MyReservationQueryService reservationQueryService) {
-        this.reservationService = reservationService;
+    public MyReservationController(MyReservationService reservationQueryService) {
         this.reservationQueryService = reservationQueryService;
     }
 
     @GetMapping("/reservations-mine")
     public ResponseEntity<List<MyReservationResponse>> getMyReservations(@AuthenticationPrincipal LoginMember loginMember) {
-        List<MyReservationResponse> myReservations = reservationService.getMyReservations(loginMember);
+        List<MyReservationResponse> myReservations = reservationQueryService.getMyReservations(loginMember);
         return ResponseEntity.ok(myReservations);
     }
